requirements.txt
Scrapy==2.11.0
markdownify==0.11.6
python-slugify==8.0.1